+++ src/steg/payload_scraper.cc
@@ -17,10 +17,10 @@ using namespace std;
 #include "http_steg_mods/file_steg.h"
 #include "http_steg_mods/jpgSteg.h"
 #include "http_steg_mods/pngSteg.h"
-#include "http_steg_mods/gifSteg.h"
-#include "http_steg_mods/swfSteg.h"
-#include "http_steg_mods/pdfSteg.h"
 #include "http_steg_mods/htmlSteg.h"
+//#include "http_steg_mods/swfSteg.h"
+//#include "http_steg_mods/pdfSteg.h"
+//#include "http_steg_mods/gifSteg.h"
 
 #include "payload_scraper.h"
 #include "base64.h"
@@ -443,7 +443,7 @@ pair<unsigned long, unsigned long> PayloadScraper::compute_capacity(string paylo
   }
   
   ssize_t capacity = cur_steg->capacity(response_head_and_body);
-  log_debug("capacity: %lu", capacity);
+  log_debug("capacity: %lu", static_cast<unsigned long>(capacity));
   if (capacity < 0){ 
     log_warn("error occurd during capacity computation");
     capacity = 0; //zero capacity files are dropped
